% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size.R
\name{adaptive_step_size}
\alias{adaptive_step_size}
\title{this function implementing the step size calculation from the adaptive proximal GD method by Yura Malisky}
\usage{
adaptive_step_size(pre_step, pre_theta, cur_L)
}
\arguments{
\item{pre_step:}{the previous step size}

\item{pre_theta:}{the previous theta parameter}

\item{cur_L:}{the current step local Lipschitz constant estimate}
}
\value{
a value represent the current step size
}
\description{
this function implementing the step size calculation from the adaptive proximal GD method by Yura Malisky
}
\keyword{internal}
