% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{Basis_generator}
\alias{Basis_generator}
\title{Generate Basis Functions for Univariate Spline}
\usage{
Basis_generator(x, N, q = 3, KnotsLocation = "quantile", knots = NULL)
}
\arguments{
\item{x}{Covariates.}

\item{N}{Number of interior knots in generating spline matrix.}

\item{q}{Degree of polynomial spline. Default is 3.}

\item{KnotsLocation}{A character string naming the way for knots locations. Default is \code{"quantile"}. The only alternative is \code{"uniform"}.}

\item{knots}{An optional vector specifying the knots to be used in constructing spline bases.}
}
\value{
A list containing the following elements:
\item{Bx0}{Non-centered spline bases for covariates.}
\item{Bx}{Centered spline bases for covariates.}
\item{BxMean}{Means of spline bases subtracted.}
\item{knots}{Knots used in constructing spline bases.}
}
\description{
This function generates basis functions for univariate splines.
}
\keyword{internal}
