% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_basis.R
\name{cbs}
\alias{cbs}
\title{Generate constant spline bases given a vector}
\usage{
cbs(x, knots, Boundary.knots = range(x))
}
\arguments{
\item{x}{Covariates.}

\item{knots}{An optional vector specifying the knots to be used in constructing spline bases.}
}
\value{
A list containing the following elements:
\item{Bx0}{Non-centered spline bases for covariates.}
\item{Bx}{Centered spline bases for covariates.}
\item{BxMean}{Means of spline bases subtracted.}
\item{knots}{Knots used in constructing spline bases.}
}
\description{
This routine generates constant spline bases given a vector
}
\keyword{internal}
