% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBIC.R
\name{compute_edf}
\alias{compute_edf}
\title{this function compute the model effective degrees of freedom}
\usage{
compute_edf(X, BQ2, P, xi_is, lambda_r)
}
\arguments{
\item{X:}{the univariate basis function matrix}

\item{BQ2:}{the bivariate matrix after QR decomposition, n by q matrix}

\item{P:}{penalty matrix (q by q matrix)}

\item{xi_is:}{the estimated slack variables from a given model}

\item{lambda_r:}{the penalty parameter for roughness}
}
\value{
a value represent the effective degrees of freedom of the model
}
\description{
this function compute the model effective degrees of freedom
}
\keyword{internal}
