% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBIC.R
\name{l_likelihood}
\alias{l_likelihood}
\title{this function helps calculate poisson log likelihood of our model with slack variables}
\usage{
l_likelihood(xi_is, Y_it, mu_t, theta_nb = 0)
}
\arguments{
\item{xi_is}{Numeric vector. The estimated slack variables (size n) for the locations.}

\item{Y_it}{Matrix. A n x t matrix of observations (counts), where columns are time points and rows are locations.}

\item{mu_t}{Numeric vector or Matrix. The mean counts. If a vector of size n is provided, it is recycled across all time points. If a matrix (n x t) is provided, it is vectorized directly.}

\item{theta_nb}{Numeric. The hyperparameter (dispersion) for the Negative Binomial family. Default is 0 (Poisson family).}
}
\value{
a value represent the log likelihood
}
\description{
this function helps calculate poisson log likelihood of our model with slack variables
}
\keyword{internal}
