% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_bps.R
\name{nb_fam}
\alias{nb_fam}
\title{define poisson family}
\usage{
nb_fam(link = "log")
}
\arguments{
\item{link:}{name of link function, default as log}
}
\value{
A group of useful functions
}
\description{
Define some function for Negative Binomial Family
}
\keyword{internal}
