% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_bps.R
\name{poisson_fam}
\alias{poisson_fam}
\title{define poisson family}
\usage{
poisson_fam(link = "log")
}
\arguments{
\item{link:}{name of link function, default as log}
}
\value{
A group of useful functions
}
\description{
Define some function for Poisson Family
}
\keyword{internal}
