% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximal_map.R
\name{proximal_map}
\alias{proximal_map}
\title{this function helps compute the gradient update during each iteration with given gradient vector}
\usage{
proximal_map(cur_gradient, lambda, h, n, w)
}
\arguments{
\item{cur_gradient:}{the current step gradient}

\item{lambda:}{the penalty parameter for xi}

\item{h:}{the step size}

\item{n:}{size of the xi vector, same as the number of observation}

\item{w:}{weight vector, same size as the xi vector}
}
\value{
a vector (same size the gradient vector) represent the proximal updates
}
\description{
this function helps compute the gradient update during each iteration with given gradient vector
}
\keyword{internal}
