% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgam_estimation.R
\name{rgam_estimation}
\alias{rgam_estimation}
\title{this function is an intermediate step of model estimation procedure, implementing the gradient descent to get parameter estimates}
\usage{
rgam_estimation(Y, X, BQ2, P, lambda1, lambda2, initial_x = NULL, w, theta_nb)
}
\arguments{
\item{Y:}{the response variable.}

\item{X:}{the covariate matrix (univariate basis).}

\item{BQ2:}{the bivariate matrix after QR decomposition.}

\item{P:}{penalty matrix.}

\item{lambda1:}{the tuning parameter for roughness.}

\item{lambda2:}{the l1 penalty parameter for slack variable.}

\item{initial_x:}{the initial value for the parameter, vector size of m (default as 0).}

\item{w:}{the weight vector for l1 penalty.}

\item{theta_nb:}{the hyperparameter for nb family (default is 0 - Poisson family).}
}
\value{
a list include parameter estimations:
\item{theta_hat}{coefficients estimated for univariate functions.}
\item{gamma_hat}{coefficients estimated for bivariate function.}
\item{xi_hat}{coefficients estimated for slack variable.}
}
\description{
this function is an intermediate step of model estimation procedure, implementing the gradient descent to get parameter estimates
}
\keyword{internal}
