% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgam_lambda_selection.R
\name{rgam_lambda_selection}
\alias{rgam_lambda_selection}
\title{Internal Function: Select Penalization Parameters for RST_GAM}
\usage{
rgam_lambda_selection(Y, X, BQ, P, lambda1, lambda2, initial_x, w, theta_nb)
}
\arguments{
\item{Y}{matrix. The response variable.}

\item{X}{matrix. The covariate matrix consisting of univariate basis.}

\item{BQ}{matrix. The matrix for bivariate basis.}

\item{P}{matrix. The smoothness penalty matrix.}

\item{lambda1}{numeric. The roughness penalty parameter for bivariate spline estimation.}

\item{lambda2}{numeric. The L1 penalty term for the slack variable.}

\item{initial_x}{numeric vector. The initial value for the parameter, default is 0.}

\item{w}{numeric vector. The weight vector for slack estimation.}

\item{theta_nb:}{the hyperparameter for nb family (default is 0 - Poisson family).}
}
\value{
A list containing the following elements:
\item{theta_hat}{coefficients estimated for univariate functions.}
\item{gamma_hat}{coefficients estimated for bivariate function.}
\item{xi_hat}{the estimated slack parameter.}
\item{r_lambda}{the selected roughness penalization parameter.}
\item{l_lambda}{the selected L1 penalization parameter.}
\item{ebic}{the final ebic value}
}
\description{
This internal function fits the global estimation model with selected penalization parameters.
}
\keyword{internal}
