% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst_gam_nb.R
\name{RST_GAM_nb}
\alias{RST_GAM_nb}
\title{Fit the Global Estimation Model (RST_GAM)}
\usage{
RST_GAM_nb(
  Y,
  X,
  BQ,
  P,
  lambda1,
  lambda2,
  initial_x,
  k,
  Ind,
  theta_range,
  fix_weight
)
}
\arguments{
\item{Y}{matrix. The response variable.}

\item{X}{matrix. The covariate matrix.}

\item{BQ}{matrix. (Bernstein basis or similar - inferred from context).}

\item{P}{matrix. (Penalty matrix or similar - inferred from context).}

\item{lambda1}{numeric. The roughness penalty parameter.}

\item{lambda2}{numeric. The L1 penalty term.}

\item{initial_x}{numeric vector. Initial values.}

\item{k}{integer. Number of data thinning folds.}

\item{Ind}{vector. Indices for the subset.}

\item{theta_range}{numeric vector. A vector of length 2 (lower, upper) specifying the search range for the NB theta parameter.}

\item{fix_weight}{vector. The weight vector for slack estimation.}
}
\value{
A list containing the following elements:
\item{theta_hat}{coefficients estimated for univariate functions.}
\item{gamma_hat}{coefficients estimated for bivariate function.}
\item{xi_hat}{the estimated slack parameter.}
\item{mu_hat}{the estimated mean parameter.}
\item{df}{degree of freedom of the model.}
\item{r_lambda}{the selected roughness penalization parameter.}
\item{l_lambda}{the selected L1 penalization parameter.}
\item{sigma_2}{estimated sigma^2.}
\item{theta_nb}{the estimated theta hyperparameter for nb family}
}
\description{
This function fits the robust ggam estimation model to account for potentail outliers using bivariate and univariate splines with Negative Binomial family.
}
\keyword{internal}
