% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rst_gam_poisson.R
\name{RST_GAM_poisson}
\alias{RST_GAM_poisson}
\title{Fit the Global Estimation Model (RST_GAM)}
\usage{
RST_GAM_poisson(
  Y,
  X,
  BQ,
  P,
  lambda1,
  lambda2,
  initial_x,
  k,
  Ind,
  theta_nb = 0,
  fix_weight = NULL,
  report_weight = FALSE
)
}
\arguments{
\item{Y}{matrix. The response variable.}

\item{X}{matrix. The covariate matrix.}

\item{BQ}{matrix. (Bernstein basis or similar - inferred from context).}

\item{P}{matrix. (Penalty matrix or similar - inferred from context).}

\item{lambda1}{numeric. The roughness penalty parameter.}

\item{lambda2}{numeric. The L1 penalty term.}

\item{initial_x}{numeric vector. Initial values.}

\item{k}{integer. Number of data thinning folds.}

\item{Ind}{vector. Indices for the subset.}

\item{theta_nb}{numeric. Hyperparameter for nb family (default 0).}

\item{fix_weight}{vector. The weight vector for slack estimation.}

\item{report_weight}{boolean. Whether to just report slack weights, used for NB model fit.}

\item{theta_nb:}{numeric. The hyperparameter for nb family (default is 0 - Poisson family).}
}
\value{
A list containing the following elements:
\item{theta_hat}{coefficients estimated for univariate functions.}
\item{gamma_hat}{coefficients estimated for bivariate function.}
\item{xi_hat}{the estimated slack parameter.}
\item{mu_hat}{the estimated mean parameter.}
\item{df}{degree of freedom of the model.}
\item{r_lambda}{the selected roughness penalization parameter.}
\item{l_lambda}{the selected L1 penalization parameter.}
\item{sigma_2}{estimated pearson sigma^2.}
}
\description{
This function fits the robust ggam estimation model to account for potentail outliers using bivariate and univariate splines with poisson family.
}
\keyword{internal}
