% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{Simulation_Data}
\alias{Simulation_Data}
\title{Generate a Simulation Dataset}
\usage{
Simulation_Data(n, t, off = 0, off_value = 100, theta_nb = 0)
}
\arguments{
\item{n}{Integer. The number of location points.}

\item{t}{Integer. The number of time points.}

\item{off}{Integer. The number of offset observations. Default is 0.}

\item{off_value}{Numeric. The offset value for observations. Default is 100.}

\item{theta_nb}{Numeric. The theta value used for Negative Binomial Family, default is 0 for Poisson Family}
}
\value{
A list containing:
\item{Y}{Matrix. The response variable.}
\item{S}{Matrix. The location information.}
\item{X}{Matrix. The covariate information.}
\item{I}{Vector. The accumulated infected cases.}
\item{Mu}{Matrix. The true mean for reponse Y}
\item{off}{Vector. The randomly generated outlier value vector}
\item{Theta_NB}{Numeric. The theta parameter used.}
}
\description{
This function generates a dataset for a simulation study, including the response variable,
location information, covariates, and accumulated infected cases.
}
\examples{
# Generate a simulation poisson dataset with 100 location points, 5 time points, no offset
sim_data <- Simulation_Data(n = 100, t = 5)

# Generate a simulation Negative Binomial dataset with 100 location points, 
# 5 time points, no offset, and theta = 5
sim_data <- Simulation_Data(n = 100, t = 5, theta_nb = 5)

# Generate a dataset with 100 locations, 5 time points, 
# 2 randomly selected outliers with value of 100
sim_data <- Simulation_Data(n = 100, t = 5, off = 2, off_value = 100)


library(dplyr)
library(ggplot2)

# plot the simulated mean parameter at each location at time 5
tibble(x = sim_data$S[,1], y = sim_data$S[,2], counts = sim_data$Mu[,5], 
offset = sim_data$off) \%>\%
  ggplot(aes(x = x, y = y, size = counts, color = offset)) +
  geom_point(alpha = 0.6) + 
  scale_radius(range = c(2,10)) +
  scale_colour_gradient(low = '#7fcdbb', high = '#253494')

}
