% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBIC.R
\name{EBIC_L}
\alias{EBIC_L}
\title{this function compute the EBIC measure of the Lasso penalty}
\usage{
EBIC_L(X, BQ2, P, xi_is, Y_it, mu_t, gamma = 0)
}
\arguments{
\item{X:}{the univariate basis function matrix}

\item{BQ2:}{the bivariate matrix after QR decomposition, n by q matrix}

\item{P:}{penalty matrix (q by q matrix)}

\item{xi_is:}{the estimated slack variables from a given model}

\item{Y_it:}{a n by t matrix representing the observations (counts) at time t (column) and location i (row) description}

\item{mu_t:}{a size n (location) vector representing the value of mean counts at time t description}

\item{gamma:}{a regulatory parameter, default as 0, note, when gamma is 0, we obtain BIC}
}
\value{
a value represent the eBIC
}
\description{
this function compute the EBIC measure of the Lasso penalty
}
\keyword{internal}
