% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_map.R
\name{gradient_map}
\alias{gradient_map}
\title{Gradient Map Function}
\usage{
gradient_map(Y, X, BQ2, P, cur_x, lambda1, theta_nb = 0)
}
\arguments{
\item{Y}{A numeric matrix.}

\item{X}{A numeric matrix.}

\item{BQ2}{A numeric matrix.}

\item{P}{A numeric matrix.}

\item{cur_x}{A numeric vector.}

\item{lambda1}{A numeric value.}

\item{theta_nb}{A numeric value (default is 0).}
}
\value{
A numeric vector of the gradient map.
}
\description{
Computes the gradient map for given parameters.
}
\keyword{internal}
