% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size.R
\name{initial_step_linesearch}
\alias{initial_step_linesearch}
\title{this function implementing a linesearch to find the initial step size}
\usage{
initial_step_linesearch(
  pre_x,
  initial_step = 10^-8,
  Y,
  X,
  BQ2,
  P,
  lambda1,
  lambda2,
  w
)
}
\arguments{
\item{pre_x:}{the previous x}

\item{initial_step:}{the starting step size, default as 10^-9}

\item{Y:}{the sample observation}

\item{X:}{the univariate basis function matrix}

\item{BQ2:}{the bivariate matrix after QR decomposition, n by q matrix}

\item{P:}{penalty matrix (q by q matrix)}

\item{lambda1:}{the penalty parameter for roughness}

\item{lambda2:}{the penalty parameter for xi}

\item{w:}{the weight vector for l1 penalty}
}
\value{
a value represent the initial step size
}
\description{
this function implementing a linesearch to find the initial step size
}
\keyword{internal}
