% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBIC.R
\name{l_likelihood}
\alias{l_likelihood}
\title{this function helps calculate poisson log likelihood of our model with slack variables}
\usage{
l_likelihood(xi_is, Y_it, mu_t)
}
\arguments{
\item{xi_is:}{the estimated slack variables from a given model}

\item{Y_it:}{a n by t matrix representing the observations (counts) at time t (column) and location i (row)}

\item{mu_t:}{a size n (location) vector representing the value of mean counts at time t}
}
\value{
a value represent the log likelihood
}
\description{
this function helps calculate poisson log likelihood of our model with slack variables
}
\keyword{internal}
